python/requirements.txt
mss==9.0.1
pytesseract==0.3.10
Pillow==10.0.0
requests==2.31.0
